# api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.9.0"
  title: Jade POC App
# basePath prefixes all resource paths
basePath: /api/0.9.0
schemes:
  # tip: remove http to make production-grade
  - http
  - https
paths:
  /device:
    # binds a127 app logic to a route
    x-swagger-router-controller: device
    get:
      description: Returns devices info
      # used as the method name of the controller
      operationId: get_device
      produces:
        - application/json
      parameters:
        - name: id
          in: query
          description: Device's ID
          required: false
          type: integer
        - name: type
          in: query
          description: Device's type
          required: false
          type: integer
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              # a pointer to a definition
              $ref: "#/definitions/Device"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/StatusResponse"
    post:
      description: Add or device
      operationId: add_device
      consumes: 
        - application/json
      parameters:
        - name: device
          in: body
          description: device to add 
          required: true
          schema:
            $ref: "#/definitions/Device"
      responses:
        "200":
          description: Success
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/StatusResponse"
    delete:
      description: Delete device
      operationId: delete_device
      parameters:
        - name: id
          description: Device's id should be removed
          in: query
          required: true
          type: integer
      responses:
        "200":
          description: Success
        default:
          description: Error
          schema:
            $ref: "#/definitions/StatusResponse"
  /user:
    # binds a127 app logic to a route
    x-swagger-router-controller: user
    get:
      description: Returns user info
      # used as the method name of the controller
      operationId: get_user
      produces: 
        - application/json
      parameters:
        - name: id
          in: query
          description: User's ID
          required: false
          type: integer
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              # a pointer to a definition
              $ref: "#/definitions/User"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/StatusResponse"
    post:
      description: Add user
      operationId: add_user
      consumes: 
        - application/json
      parameters:
        - name: user
          in: body
          description: user to add 
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Success
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/StatusResponse"
    delete:
      description: Delete user
      operationId: delete_user
      parameters:
        - name: id
          description: User's id should be removed
          in: query
          required: true
          type: integer
      responses:
        "200":
          description: Success
        default:
          description: Error
          schema:
            $ref: "#/definitions/StatusResponse"
            
# complex objects have schema definitions
definitions:
  Device:
    required:
      - id
      - type
    properties:
      id:
        type: integer
      type:
        type: integer
      description:
        type: string
      ip:
        type: string
  User:
    required:
      - id
      - name
      - profile
    properties:
      id:
        type: integer
      name:
        type: string
      profile:
        type: string
  StatusResponse:
    required:
      - message
      - code
    properties:
      message:
        type: string
      code:
        type: integer